Add tests for subscription callback
